fix: rename ThreadChannel.js folder
