+++ packages/discord.js/DD2480-docs/functions/ThreadChannel.js/ThreadChannel.js
